+++ Framework/Binding/RootViewFactory.cs
@@ -42,6 +42,7 @@ internal class RootViewFactory(IEnumerable<IViewFactory> addonFactories, CommonS
             "spacer" => new Spacer(),
             "tab" => new Tab(),
             "textinput" => new TextInput(),
+            "item-icon" => new ItemIcon(),
             _ => addonFactories.FirstOrDefault(factory => factory.SupportsTag(tagName))?.CreateView(tagName)
                 ?? throw new ArgumentException($"Unsupported view type: {tagName}", nameof(tagName)),
         };